refactor(core): lift prepare step into seperate precompute package
