+++ alvoc/core/mutations/utils.py
@@ -1,6 +1,21 @@
 from pysam import PileupColumn
 from alvoc.core.mutations.convert import aa
 
+def parse_mutation(mut: str, genes: dict, seq: str):
+    """Parse mutation to handle potential multiple SNVs.
+
+    Args:
+        mut : Mutation string, which could include multiple SNVs.
+        genes : Dictionary of genes with start and end positions.
+        seq : The nucleotide sequence.
+
+    Returns:
+        list: A list of parsed SNVs.
+    """
+    muts = aa(mut, genes, seq) if ':' in mut else [mut]
+    return [parse_snv(m) for m in muts]
+
+
 def parse_snv(snv: str):
     """Parse a single nucleotide variant (SNV).
 
@@ -26,22 +41,6 @@ def snv_name(snv: tuple):
         str: A string representation of the SNV.
     """
     return '{}{}{}'.format(*snv)
-
-
-def parse_mutation(mut: str, genes: dict, seq: str):
-    """Parse mutation to handle potential multiple SNVs.
-
-    Args:
-        mut : Mutation string, which could include multiple SNVs.
-        genes : Dictionary of genes with start and end positions.
-        seq : The nucleotide sequence.
-
-    Returns:
-        list: A list of parsed SNVs.
-    """
-    muts = aa(mut, genes, seq) if ':' in mut else [mut]
-    return [parse_snv(m) for m in muts]
-
 
 def mut_in_col(pileupcolumn: PileupColumn, mut : str):
     """Count the occurrences and non-occurrences of a mutation in a pileup column.
